# platform/contracts/openapi/public-api.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Unified Software Commercialization Platform Public API",
    "version": "1.0.0-draft.1",
    "description": "Contract-first public client and admin API. Product, Application, Tenant and User contexts are resolved from verified sessions; identifiers supplied by clients are never authorization facts."
  },
  "servers": [{ "url": "/", "description": "Current environment" }],
...
  "paths": {
...
    "/api/v1/admin/products": {
      "get": {
        "tags": ["Product"], "operationId": "listProducts", "summary": "List products visible to the administrator",
        "security": [{ "AdminAccessCookie": [] }, { "AdminBearer": [] }],
        "parameters": [{ "$ref": "#/components/parameters/RequestId" }],
        "responses": { "200": { "$ref": "#/components/responses/ProductList" }, "default": { "$ref": "#/components/responses/Error" } }
      },
      "post": {
        "tags": ["Product"], "operationId": "createProduct", "summary": "Create a product and its official tenant",
        "security": [{ "AdminAccessCookie": [] }, { "AdminBearer": [] }],
        "parameters": [{ "$ref": "#/components/parameters/RequestId" }, { "$ref": "#/components/parameters/IdempotencyKey" }, { "$ref": "#/components/parameters/AdminCsrfToken" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateProductRequest" } } } },
        "responses": { "201": { "$ref": "#/components/responses/Product" }, "default": { "$ref": "#/components/responses/Error" } }
      }
    },
...
    "/api/v1/admin/audit/events": {
      "get": {
        "tags": ["Audit"], "operationId": "listAuditEvents", "summary": "Search redacted audit events inside the authorized scope",
        "security": [{ "AdminAccessCookie": [] }, { "AdminBearer": [] }],
        "parameters": [{ "$ref": "#/components/parameters/RequestId" }, { "$ref": "#/components/parameters/Cursor" }, { "name": "limit", "in": "query", "description": "Bounded page size implemented by the audit query service.", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 } }, { "name": "trace_id", "in": "query", "schema": { "type": "string", "maxLength": 256 } }],
        "responses": { "200": { "$ref": "#/components/responses/AuditPage" }, "default": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/api/v1/admin/audit/events/{audit_id}": {
      "parameters": [{ "$ref": "#/components/parameters/AuditId" }],
      "get": {
        "tags": ["Audit"], "operationId": "getAuditEvent", "summary": "Read one redacted audit event after authorizing its recorded scope",
        "description": "A missing event and an event outside the administrator's current audit.read scope return the same not-found envelope. Callers may retry briefly while a business outbox is delivering the event.",
        "x-required-permission": "audit.read", "x-scope-derived-from-resource": true,
        "security": [{ "AdminAccessCookie": [] }, { "AdminBearer": [] }],
        "parameters": [{ "$ref": "#/components/parameters/RequestId" }],
        "responses": { "200": { "$ref": "#/components/responses/AuditEvent" }, "default": { "$ref": "#/components/responses/Error" } }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ClientSessionBearer": { "type": "http", "scheme": "bearer", "bearerFormat": "opaque", "description": "Short-lived client session containing server-resolved ProductContext, ApplicationContext and TenantContext." },
      "UserBearer": { "type": "http", "scheme": "bearer", "bearerFormat": "opaque", "description": "User session bound to trusted Product/Application and server-resolved Tenant context." },
      "HostedSessionCookie": { "type": "apiKey", "in": "cookie", "name": "__Host-platform_hosted_session", "description": "Short-lived Hosted UI browser session. Server sets Secure, HttpOnly, SameSite=Lax, Path=/ and no Domain. Unsafe methods also require exact Hosted Origin and X-CSRF-Token." },
      "AdminAccessCookie": { "type": "apiKey", "in": "cookie", "name": "__Host-platform_admin_access", "description": "Default browser administrator access. Server sets Secure, HttpOnly, SameSite=Strict, Path=/ and no Domain. Unsafe methods also require exact Origin and X-CSRF-Token." },
      "AdminRefreshCookie": { "type": "apiKey", "in": "cookie", "name": "__Secure-platform_admin_refresh", "description": "Rotating browser refresh proof. Server sets Secure, HttpOnly, SameSite=Strict, Path=/api/v1/admin/auth and no Domain; accepted only by admin refresh/logout." },
      "AdminBearer": { "type": "http", "scheme": "bearer", "bearerFormat": "opaque", "description": "Optional non-browser administrator transport explicitly issued to controlled CLI/automation; every target product/tenant is re-authorized by Access Control." },
      "DeveloperKey": { "type": "http", "scheme": "bearer", "bearerFormat": "developer-key", "description": "Restricted developer key. The key resolves Product/Tenant/scope server-side and cannot select arbitrary contexts." }
    },
    "parameters": {
      "RequestId": { "name": "X-Request-Id", "in": "header", "required": false, "description": "Optional caller correlation ID; not an authorization or idempotency fact.", "schema": { "type": "string", "minLength": 8, "maxLength": 128 } },
      "IdempotencyKey": { "name": "Idempotency-Key", "in": "header", "required": true, "description": "Opaque operation-scoped idempotency key. Reuse with a different payload returns conflict.", "schema": { "type": "string", "minLength": 16, "maxLength": 128 } },
      "AdminCsrfToken": { "name": "X-CSRF-Token", "in": "header", "required": false, "description": "Required with exact Origin for unsafe requests authenticated by an administrator access Cookie, except refresh which recovers after access expiry using exact Origin plus SameSite refresh Cookie. Bound to the server session and rotated with it. Omit for Admin Bearer.", "schema": { "type": "string", "minLength": 32, "maxLength": 256 } },
...
      "Cursor": { "name": "cursor", "in": "query", "schema": { "type": "string", "maxLength": 512 } },
...
      "AuditId": { "name": "audit_id", "in": "path", "required": true, "description": "Stable audit event identifier; authorization is derived from the event's recorded scope.", "schema": { "$ref": "#/components/schemas/Identifier" } },
...
    },
    "headers": {
      "RequestId": { "description": "Server request/correlation ID.", "schema": { "type": "string" } }
    },
    "responses": {
      "Error": { "description": "Stable safe error", "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } }, "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
...
      "Product": { "$ref": "#/components/responses/JsonProduct" },
      "JsonProduct": { "description": "Product", "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Product" } } } },
...
      "ProductList": { "description": "Authorized products", "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } }, "content": { "application/json": { "schema": { "type": "object", "required": ["items"], "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/ProductSummary" } } }, "additionalProperties": false } } } },
...
      "AuditPage": { "description": "Redacted audit events", "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuditPage" } } } },
      "AuditEvent": { "description": "One redacted audit event", "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuditEvent" } } } }
    },
    "schemas": {
      "Identifier": { "type": "string", "minLength": 1, "maxLength": 128, "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$" },
      "StableCode": { "type": "string", "minLength": 3, "maxLength": 64, "pattern": "^[a-z][a-z0-9]*(?:[._-][a-z0-9]+)*$" },
...
      "Timestamp": { "type": "string", "format": "date-time" },
...
      "ErrorEnvelope": { "type": "object", "required": ["type", "title", "status", "code", "request_id", "retryable"], "properties": { "type": { "type": "string", "format": "uri-reference" }, "title": { "type": "string" }, "status": { "type": "integer", "minimum": 400, "maximum": 599 }, "code": { "type": "string", "pattern": "^[A-Za-z0-9_.-]+$" }, "detail": { "type": "string" }, "request_id": { "type": "string" }, "retryable": { "type": "boolean" }, "retry_after_seconds": { "type": "integer", "minimum": 0 }, "field_errors": { "type": "array", "items": { "type": "object", "required": ["field", "code"], "properties": { "field": { "type": "string" }, "code": { "type": "string" }, "message": { "type": "string" } } } } } },
...
      "CreateProductRequest": { "type": "object", "required": ["code", "name", "status"], "properties": { "code": { "$ref": "#/components/schemas/StableCode" }, "name": { "type": "string", "minLength": 1, "maxLength": 120 }, "status": { "type": "string", "enum": ["active", "suspended"] } }, "additionalProperties": false },
      "Product": { "type": "object", "required": ["product_id", "code", "name", "status", "provisioning_state", "official_tenant_id", "context_version", "created_at", "updated_at", "audit_id"], "properties": { "product_id": { "$ref": "#/components/schemas/Identifier" }, "code": { "$ref": "#/components/schemas/StableCode" }, "name": { "type": "string" }, "status": { "type": "string", "enum": ["active", "suspended"] }, "provisioning_state": { "type": "string", "const": "ready" }, "official_tenant_id": { "$ref": "#/components/schemas/Identifier" }, "context_version": { "type": "integer", "minimum": 1 }, "created_at": { "$ref": "#/components/schemas/Timestamp" }, "updated_at": { "$ref": "#/components/schemas/Timestamp" }, "audit_id": { "$ref": "#/components/schemas/Identifier" } }, "additionalProperties": false },
      "ProductSummary": { "type": "object", "required": ["product_id", "code", "name", "status", "provisioning_state", "context_version", "created_at", "updated_at"], "properties": { "product_id": { "$ref": "#/components/schemas/Identifier" }, "code": { "$ref": "#/components/schemas/StableCode" }, "name": { "type": "string" }, "status": { "type": "string", "enum": ["active", "suspended"] }, "provisioning_state": { "type": "string", "enum": ["pending", "ready", "failed"] }, "official_tenant_id": { "type": ["string", "null"], "maxLength": 128 }, "context_version": { "type": "integer", "minimum": 1 }, "created_at": { "$ref": "#/components/schemas/Timestamp" }, "updated_at": { "$ref": "#/components/schemas/Timestamp" } }, "additionalProperties": false },
...
      "AuditEvent": { "type": "object", "required": ["audit_id", "occurred_at", "actor_id", "permission", "scope_type", "action", "target_type", "target_id", "result", "trace_id"], "properties": { "audit_id": { "$ref": "#/components/schemas/Identifier" }, "occurred_at": { "$ref": "#/components/schemas/Timestamp" }, "actor_id": { "$ref": "#/components/schemas/Identifier" }, "permission": { "type": "string" }, "scope_type": { "type": "string", "enum": ["platform", "product", "tenant"] }, "scope_id": { "$ref": "#/components/schemas/Identifier" }, "action": { "type": "string" }, "target_type": { "type": "string" }, "target_id": { "$ref": "#/components/schemas/Identifier" }, "result": { "type": "string" }, "reason_code": { "type": ["string", "null"] }, "trace_id": { "type": "string" }, "redacted_summary": { "type": "object", "additionalProperties": true } } },
      "AuditPage": { "type": "object", "required": ["items"], "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/AuditEvent" } }, "next_cursor": { "type": ["string", "null"] } } }
    }
  }
}